+++ examples/clerk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # minimal-auth-clerk
 
+## 0.0.74
+
+### Patch Changes
+
+- Updated dependencies [fe6f561]
+  - jazz-tools@0.13.5
+  - jazz-react@0.13.5
+  - jazz-react-auth-clerk@0.13.5
+
 ## 0.0.73
 
 ### Patch Changes