+++ examples/multi-cursors/CHANGELOG.md
@@ -1,5 +1,13 @@
 # multi-cursors
 
+## 0.0.68
+
+### Patch Changes
+
+- Updated dependencies [fe6f561]
+  - jazz-tools@0.13.5
+  - jazz-react@0.13.5
+
 ## 0.0.67
 
 ### Patch Changes