+++ examples/music-player/CHANGELOG.md
@@ -1,5 +1,22 @@
 # jazz-example-musicplayer
 
+## 0.0.97
+
+### Patch Changes
+
+- Updated dependencies [c320615]
+  - jazz-inspector@0.13.6
+
+## 0.0.96
+
+### Patch Changes
+
+- Updated dependencies [08ae9b2]
+- Updated dependencies [fe6f561]
+  - jazz-inspector@0.13.5
+  - jazz-tools@0.13.5
+  - jazz-react@0.13.5
+
 ## 0.0.95
 
 ### Patch Changes