+++ examples/passphrase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # passphrase
 
+## 0.0.70
+
+### Patch Changes
+
+- Updated dependencies [fe6f561]
+  - jazz-tools@0.13.5
+  - jazz-react@0.13.5
+
 ## 0.0.69
 
 ### Patch Changes