+++ examples/todo-vue/CHANGELOG.md
@@ -1,5 +1,14 @@
 # todo-vue
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies [fe6f561]
+  - jazz-tools@0.13.5
+  - jazz-browser@0.13.5
+  - jazz-vue@0.13.5
+
 ## 0.0.75
 
 ### Patch Changes