+++ packages/jazz-react-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-react-core
 
+## 0.13.5
+
+### Patch Changes
+
+- Updated dependencies [e090b39]
+- Updated dependencies [fe6f561]
+  - cojson@0.13.5
+  - jazz-tools@0.13.5
+
 ## 0.13.4
 
 ### Patch Changes